+++ src/task_manager/core/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 from django.contrib.auth.admin import UserAdmin
-from user.models import User
+from core.models import User
 # Register your models here.
 
 admin.site.register(User, UserAdmin)
